cloud-sql-tuning-agent/requirements.txt
pandas
openai
python-dateutil
sqlparse
psycopg2-binary
mysql-connector-python
tabulate

